feat(urls): switch to params for stable URLs
In the course of working on stable URLs I discovered that loading the application state from the pathname would greatly increase complexity. I'd need to introduce either a server and/or a router. Since I'd like to avoid that complexity as long as possible, I've opted to use searchParams instead.