task(state): create initial application state
This change sets the stage for state management in the application
by creating default data, including options for all fields as well
as some meta data about the fields. At the moment, that's just a key
indicating whether the field is locked (defaults to false).

Later on this inital state will be loaded into local storage if
state hasn't already been loaded. Local storage will be the source
of truth for the state of the application from then on.